python-backend/requirements.txt
numpy==1.24.3
opencv-python==4.8.0.76
tensorflow==2.14.0
face-recognition==1.4.0
flask==3.0.0
flask-cors==4.0.0
Pillow==10.0.0
scikit-image==0.21.0
scikit-learn==1.3.0
python-dotenv==1.0.0
